refactor(dispatch): extract BuildDispatch, the final launching lane
Move the spec-dispatch lane out of DispatchOperations into a dedicated
BuildDispatch collaborator: dispatch, claimAndPrepare, branch checkout
(checkoutBranch/baseBranch/fetchLatestBase), spec resolution, repo-overlap
guard, run reservation, and the lane-local publish/exec/seedRoomDelivery
helpers. DispatchOperations keeps only the field, constructor wiring, and a
thin dispatch() delegator, dropping from ~913 to 458 lines — now a facade
over the four launching lanes (adhoc, room-wake, invite, build) and the
three seams (LaunchAdmission, RunReservation, RunLauncher).

Behavior-preserving: the full dispatch/lane suite is unchanged and green.
resolveSpec/SpecResolution moved with the lane; the resolve-spec test now
references BuildDispatch. BuildDispatchCoverageTest locks the lane's edges
through real dispatches — the no-agent refusal, forking off a pinned base
branch, a shell failure mapping to COMMAND_FAILED, and room-message
delivery ledgering at launch — for honest 100% coverage, no floor.
